- Introduce belt elongation formula
- Refactor vertimag-configuration.json format

Related work items: #2029

+++ VerticalWarehouses/Machine Automation Service/Ferretto.VW.MAS.AutomationService/Models/VerticalOffsetProcedureParameters.cs
@@ -6,9 +6,9 @@
 
         public int ReferenceCellId { get; set; }
 
-        public decimal StepValue { get; set; }
+        public double StepValue { get; set; }
 
-        public decimal VerticalOffset { get; set; }
+        public double VerticalOffset { get; set; }
 
         #endregion
     }
